python_version/requirements.txt
chromadb==0.3.21
langchain==0.0.146
python-dotenv==1.0.0
GitPython==3.1.31